don't split the kernel class
